% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{$,DRM-method}
\alias{$,DRM-method}
\title{Allow the `$` accessor with S4 objects}
\usage{
\S4method{$}{DRM}(x, name)
}
\arguments{
\item{x}{Object of class 'DRM'.}

\item{name}{a valid name (an element of \code{slotNames(x)}).}
}
\description{
While S4 slots are naturally available for DRM objects, most users are more familiar with operator$, and it may be more convenient.
}
\seealso{
\code{\link[mdrm:names-DRM-method]{names}}
}
