% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{$,DRM_-method}
\alias{$,DRM_-method}
\title{Allow the familiar `$` accessor with S4 objects}
\usage{
\S4method{$}{DRM_}(x, name)
}
\arguments{
\item{x}{Object of class 'DRM'}

\item{name}{a valid name (a member of \code{names(x)})}
}
\description{
While S4 slots are naturally available for DRM_ objects, most users are more familiar with operator$, and it may be more convenient.
}
\seealso{
\code{\link[mdrm:names-DRM_-method]{names}}
}
