% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{coefficients,DRM-method}
\alias{coefficients,DRM-method}
\title{Get model coefficients}
\usage{
\S4method{coefficients}{DRM}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}

\item{...}{a list; the named argument 'which' is parsed and may take values 'alpha', 'beta', or 'Fhat'.
In this case, only that subset of parameters is returned.
otherwise (by default), returns a list with all three.}
}
\value{
A matrix or vector (if argument \code{which} is used), or a list with components \code{alpha}
and \code{beta}.
}
\description{
It is probably just as easy to use the slot names directly, but for consistency this method is available, too
}
