% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{class}
\name{DRM-class}
\alias{DRM-class}
\alias{DRM}
\title{An S4 class for fitting density ratio model and summarizing results}
\description{
An S4 class for fitting density ratio model and summarizing results
}
\section{Slots}{

\describe{
\item{\code{X}}{the regressors}

\item{\code{Y}}{the outcome(s)}

\item{\code{U}}{unique values of Y}

\item{\code{nUniq}}{multiplicities of unique values of Y}

\item{\code{beta}}{regression parameters}

\item{\code{Fhat}}{jumps sizes of functional parameter}

\item{\code{N}}{sample size}

\item{\code{resid}}{model residuals}

\item{\code{sdbeta}}{standard errors of beta}

\item{\code{sdF}}{standard errors of Fhat}

\item{\code{covHat}}{covariance matrix of all parameters}

\item{\code{p.value}}{p-values for beta}

\item{\code{logLik}}{log-likelihood of MLE}

\item{\code{iterations}}{number of iterations incurred during the fitting process}

\item{\code{alpha}}{parameters which are softmax-transformed to attain Fhat}

\item{\code{betaC}}{beta scaled by the sample covariance of residuals}

\item{\code{user.call}}{used in displayed output; records the formula used to generate X and Y}

\item{\code{na_omit}}{indices of any values which were omitted due to missingness}

\item{\code{zc}}{records indices of any parameters which were requested to fix at zero during fitting}
}}

