% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\name{drmBootstrap}
\alias{drmBootstrap}
\title{Bootstrap estimation of residuals' (co)variance}
\usage{
drmBootstrap(formula, data, subset, na.action, B, summarize = FALSE,
  contrasts = NULL, fitOptions = list())
}
\arguments{
\item{formula}{Model formula specifying response(s) and covariates}

\item{data}{See corresponding documentation for \code{\link{drm}}}

\item{subset}{See corresponding documentation for \code{\link{drm}}}

\item{na.action}{See corresponding documentation for \code{\link{drm}}}

\item{B}{number of bootstrap replications}

\item{summarize}{Should this function return the entire length-\code{B} sequences of fitted \eqn{\hat{\beta}}{\beta} and \eqn{\hat{\Sigma}}{\Sigma}, or just the empirical standard deviation of \eqn{\beta\Sigma}?}

\item{contrasts}{See corresponding documentation for \code{\link{drm}}}

\item{fitOptions}{See corresponding documentation for \code{\link{drm}}}
}
\value{
If \code{summarize=TRUE}, a matrix of standard deviations. Otherwise, two lists of length \code{B}: the fitted \eqn{\beta} and corresponding \eqn{\Sigma}
}
\description{
In the case of scaling the estimated coefficients \eqn{\beta} by the residual covariance, estimate
the standard errors of the scaled parameters via non-parametric bootstrap
}
\details{
Word of caution: if the data set and/or requested number of bootstrap samples is very large, we may overflow RAM when summarize is
FALSE since storage is on the order \code{B*N*(ncol(X)+ncol(Y))}.  When summarize = TRUE, the factor of \code{N} disappears.
}
\seealso{
\code{\link{drm}}
}
