% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{fitted,DRM_-method}
\alias{fitted,DRM_-method}
\title{Extract Model Fitted Values}
\usage{
\S4method{fitted}{DRM_}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}
}
\value{
The fitted values of the response \code{Y}.
}
\description{
Identical to \code{object@Y - resid(object)}.
}
