% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{logLik,DRM-method}
\alias{logLik,DRM-method}
\title{Extract log-likelihood}
\usage{
\S4method{logLik}{DRM}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}

\item{...}{extra arguments to satisfy method dispatch; are ignored.}
}
\value{
The log-likelihood at the MLE for this object.
}
\description{
Provides an S3-generic method for getting log-likelihood from objects of DRM class
}
