% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{lr_test,DRM_-method}
\alias{lr_test,DRM_-method}
\title{Likelihood ratio test for drm object}
\usage{
\S4method{lr_test}{DRM_}(obj, tstMat, literal = FALSE)
}
\arguments{
\item{obj}{A fitted drm object}

\item{tstMat}{The parameter values for \eqn{\beta}}

\item{literal}{A boolean indicating whether the provided values of tstMat should be used verbatim or as indicators of which elements of
the MLE to set to zero.}
}
\description{
Method for drm class objects to conduct a likelihood ratio test.  Uses the likelihood function of the density ratio model to conduct inference on \eqn{\beta}
}
\seealso{
\code{\link[mdrm:summary-DRM_-method]{summary}}
}
