% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrmPackageDoc.R
\docType{package}
\name{MDRM}
\alias{MDRM}
\alias{MDRM-package}
\title{MDRM: a package for fitting a semi-parametric \emph{density ratio model}}
\description{
This package's principal function (for now) is \code{drm}, which fits a model
for data which follows an exponential family distribution.
}
\section{Functions}{

drm 
lr_test
summary 
plot 
print
vcov
drmBoot
}

