% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{names,DRM-method}
\alias{names,DRM-method}
\title{List the names (attributes) of a DRM object}
\usage{
\S4method{names}{DRM}(x)
}
\arguments{
\item{x}{Object of class 'DRM'}
}
\value{
A list of object names.
}
\description{
Equivalent to \code{names(d)} when d has class data.frame, lm, etc (i.e. when it is a list).
Use x$name to see the contents of the named attribute; for example x$sdbeta 
will return the matrix of estimated standard deviations of the model's \eqn{\beta}{beta} coefficients.
}
\seealso{
\code{\link{drm}}, \code{\link[mdrm:summary-DRM-method]{summary}}
}
