% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{nobs,DRM-method}
\alias{nobs,DRM-method}
\title{Get the sample size used in model fitting}
\usage{
\S4method{nobs}{DRM}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}

\item{...}{extra arguments to satisfy method dispatch; are ignored.}
}
\value{
The sample size \eqn{N} of data used to fit this model (excludes missing observations).
}
\description{
This is mostly for interface with the sandwich package. Just returns \code{object@N}.
}
