% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\docType{methods}
\name{plot,DRM_-method}
\alias{plot,DRM_-method}
\title{plot the empirical baseline distribution function.}
\usage{
\S4method{plot}{DRM_}(x, y, ...)
}
\arguments{
\item{x}{an object of class 'DRM', created by calling \code{\link{drm}}}
}
\value{
nothing; called for its side effect of creating a plot
}
\description{
\samp{plot{x}} plots the baseline CDF based on the estimated jump sizes \eqn{F}, where \eqn{x} is an objet of class \code{DRM}
}
\details{
The number of distinct jumps equals the number of unique \eqn{Y} observations (calculated row-wise if Y is a matrix like \code{unique(Y)}).
For univariate Y, 95% confidence bands are included.  Note that, given \eqn{m} unique outcomes, model identifiability requires \eqn{\sum_{j=1}^{m}F_j=1}
so that there are \eqn{m-1} parameters. We set \eqn{F_{m}\equiv 1-\sum_{j=1}^{m-1}F_j}.

Note that plotting is only implemented for univariate and bivariate outcomes.
}
\seealso{
\code{\link{drm}}
}
