% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{predict,DRM-method}
\alias{predict,DRM-method}
\title{Predict method for mdrm object}
\usage{
\S4method{predict}{DRM}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted MDRM object (as returned from \code{mdrm})}

\item{newdata}{A \code{data.frame} or \code{matrix} whose model matrix has the same columns as
the one used to fit \code{object}}

\item{...}{other arguments; currently ignored}
}
\value{
matrix of predicted Y values
}
\description{
Predict values based on new data or the data used to fit the model
}
