% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{print,DRM-method}
\alias{print,DRM-method}
\title{Print an object of class 'DRM'}
\usage{
\S4method{print}{DRM}(x)
}
\arguments{
\item{x}{an object returned from calling \code{drm}}
}
\value{
Information printed to screen.
}
\description{
prints a brief summary of an object retured from calling \code{drm}.  See \code{\link[mdrm:names-DRM-method]{names}} for a more thorough object description.
}
