% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{summary,DRM_-method}
\alias{summary,DRM_-method}
\title{Summary table of model coefficients}
\usage{
\S4method{summary}{DRM_}(object)
}
\arguments{
\item{object}{an object of class 'DRM'}
}
\value{
A summary table of Wald tests.
}
\description{
Coefficient estimates, standard deviations, z-scores and p-values for the \emph{parametric} parameters \eqn{\hat{\beta}}
}
\details{
Asymptotically, the non-parametric MLE coefficients \deqn{\hat{\beta}}{beta-hat} follow a zero-mean normal distribution.
The summary table provides the p-values for a two-sided Wald test.
}
\seealso{
\code{\link[mdrm:lr_test-method]{lr_test}} for testing multiple coefficients via a likelihood-ratio test.
}
