% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{vcov,DRM_-method}
\alias{vcov,DRM_-method}
\title{Get the covariance matrix of model parameters}
\usage{
\S4method{vcov}{DRM_}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}

\item{...}{more arguments (for consistency with the S3 generic function). They're ignored.}
}
\value{
The variance-covariance matrix of parameter estimates. The top left and bottom right blocks contain the covariance
of the jump sizes \eqn{\hat{F}} and the mean parameters \eqn{\hat{\beta}}, respectively.
}
\description{
This is just a wrapper which returns object@covHat, to keep a consistent interface.
}
