% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2u.R
\name{Y2U}
\alias{Y2U}
\title{Partially discretize a 1- or 2- dimensional outcome}
\usage{
Y2U(Y, p = 0.5, plot = FALSE)
}
\arguments{
\item{Y}{a matrix or data.frame whose mode is "numeric" or "logical" (for factor/character 
data, coerce first with \code{as.integer(as.factor(x))}).}

\item{p}{The scaling exponent. If at least one column of Y is continuous in the sense 
that all of its values are distinct, then the goal is to return a matrix with 
approximately \code{nrow(Y)^p} distinct values. Since the number of distinct values 
can only decrease, it's required that \eqn{0<p<1}.
Note that the value of \code{p} entails a target and not a guarantee.}

\item{plot}{In the bivariate case, provides a plot as a side effect superimposing 
the raw and binned data so that the accuracy of the binning can be inspected.}
}
\value{
A matrix \code{U} whose dimensions match those of \code{Y}
}
\description{
\samp{Y2U} can be used to bin continuous data, which allows fitting non-parametric 
 models with "infinite-dimensional" (functional) parameters when using the raw data would 
 prove computationally infeasible, as there is one parameter per distinct observation.
}
\details{
There are three cases handled here: for a single, univariate outcome the returned 
value has O(n^p) distinct values if the input Y has n distinct values. For continuous, 
bivariate data, the \code{hexbin::hexbin} is used to bin values in the plane. 
When one margin is continuous and another discrete, binning of the continuous variable
only occurs within the same value of the discrete variable. Note that the model-based 
variance estimator using binned data will tend to be biased toward zero; using a 
bootstrap estimate of variance may yield more accurate coverage probability.
}
\examples{
\donttest{
YY = matrix(rnorm(8000),ncol=2)
U1 = Y2U(YY[,1])
U2 = Y2U(YY,0.4)
c(length(unique(U1)),nrow(unique(U2)))
cor(U1,YY[,1])
c(cor(U2[,1],YY[,1]),cor(U2[,2],YY[,2]))
}

}
