% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{logLik,DRM_-method}
\alias{logLik,DRM_-method}
\title{Extract log-likelihood}
\usage{
\S4method{logLik}{DRM_}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}

\item{...}{extra arguments to satisfy method dispatch; are ignored}
}
\description{
Provides an S3-generic method for getting log-likelihood from objects of DRM class
}
