% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{lr_test}
\alias{lr_test}
\alias{lr_test,character,ANY-method}
\title{Likelihood ratio test}
\usage{
lr_test(obj, tstMat, literal)
}
\arguments{
\item{obj}{An object returned by \code{drm()}.}

\item{tstMat}{The parameter values corresponding to the null hypothesis (assuming that the MLE is the alternative)}

\item{literal}{A boolean indicating whether the provided values of tstMat should be used verbatim or as 
indicators of which elements of the MLE to set to zero.}
}
\value{
A list detailing the results of the likelihood ratio test.
}
\description{
Perform likelihood ratio test for mean parameters in density ratio model.
}
\details{
This is a generic function which is implemented as an S4 method for the DRM class

Calculates the result of a \eqn{\chi^2}{chi-squared} test for \eqn{H_0:}{H0:} specified elements of 
\eqn{\beta}{beta} are zero vs. the alternative in which the likelihood is evaluated at 
\eqn{\hat{\beta}_{MLE}}{beta_MLE}.
}
