% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmClass.R
\docType{methods}
\name{nobs,DRM_-method}
\alias{nobs,DRM_-method}
\title{Get the sample size used in model fitting}
\usage{
\S4method{nobs}{DRM_}(object, ...)
}
\arguments{
\item{object}{a fitted drm object}

\item{...}{more arguments (for consistency with the S3 generic function). It is ignored.}
}
\value{
The sample size N of this model
}
\description{
This is mostly for interface with the sandwich package. Just returns \code{object@N}
}
